+++ src/main/java/de/htwsaar/prog3/carrental/gui/CarTableView.java
@@ -2,10 +2,10 @@ package de.htwsaar.prog3.carrental.gui;
 
 import java.util.Locale;
 
-import de.htwsaar.prog3.carrental.i18n.I18nComponentsUtil;
-import de.htwsaar.prog3.carrental.i18n.I18nStringsUtil;
-import de.htwsaar.prog3.carrental.i18n.I18nUtil;
 import de.htwsaar.prog3.carrental.util.EntityManagerUtil;
+import de.htwsaar.prog3.carrental.util.I18nComponentsUtil;
+import de.htwsaar.prog3.carrental.util.I18nStringsUtil;
+import de.htwsaar.prog3.carrental.util.I18nUtil;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
