+++ src/main/java/de/htwsaar/prog3/carrental/util/GUIDialogUtil.java
@@ -1,6 +1,5 @@
 package de.htwsaar.prog3.carrental.util;
 
-import de.htwsaar.prog3.carrental.i18n.I18nComponentsUtil;
 import javafx.scene.control.Alert;
 import javafx.scene.control.Alert.AlertType;
